Move Metric class into its own file
To avoid import cycle issues
